--- Projects/Features/CommonFeature/Sources/PlayState/PlayerMovement.swift
@@ -1,16 +1,0 @@
-//
-//  PlayerMovement.swift
-//  CommonFeature
-//
-//  Created by KTH on 2023/03/30.
-//  Copyright © 2023 yongbeomkwak. All rights reserved.
-//
-
-import Foundation
-import DomainModule
-
-public enum PlayerMovement: Equatable {
-    case full
-    case mini
-    case close
-}
